+++ lib/startback/audit/trailer.rb
@@ -52,6 +52,7 @@ module Startback
 
         def call(severity, time, progname, msg)
           if msg[:error] && msg[:error].respond_to?(:message, true)
+            msg[:backtrace] = msg[:error].backtrace[0..25] if severity == "FATAL"
             msg[:error] = msg[:error].message
           end
           {